+++ packages/web/index.html
@@ -3,41 +3,7 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>VS Code JSON Schema Form</title>
-    <style>
-        body {
-            margin: 0;
-            padding: 20px;
-            background-color: #1e1e1e;
-            color: #cccccc;
-            font-family: system-ui, -apple-system, sans-serif;
-        }
-        .field-container {
-            margin-bottom: 1rem;
-        }
-        .field-container label {
-            display: block;
-            margin-bottom: 0.5rem;
-        }
-        .description {
-            display: block;
-            margin-top: 0.25rem;
-            color: #888888;
-        }
-        .markdown-output {
-            margin-top: 2rem;
-            padding: 1rem;
-            background-color: #2d2d2d;
-            border-radius: 4px;
-        }
-        /* VS Code-like theme for the form elements */
-        vscode-text-field,
-        vscode-text-area,
-        vscode-dropdown {
-            margin-bottom: 8px;
-            width: 100%;
-        }
-    </style>
+    <title>GenAIScript Script Runner</title>
 </head>
 <body>
     <div id="root"></div>
