+++ packages/web/src/index.tsx
@@ -2,7 +2,7 @@
   // src/index.tsx
   import React from 'react';
   import { createRoot } from 'react-dom/client';
-  import { JsonSchemaForm } from './components/JsonSchemaForm';
+  import JsonSchemaForm from './components/JSONSchemaForm';
   import { sampleSchema } from './schema';
   
   const rootElement = document.getElementById('root');
